backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.5.0